+++ components/Provider.tsx
@@ -4,7 +4,7 @@ import { SessionProvider } from "next-auth/react";
 
 interface providerProps {
   children: any;
-  session: any;
+  session?: any;
 }
 
 const Provider = ({ children, session }: providerProps) => {
