+++ next.config.js
@@ -17,13 +17,13 @@ const nextConfig = {
       },
     ],
   },
-  // webpack(config) {
-  //   config.experiments = {
-  //     ...config.experiments,
-  //     topLevelAwait: true,
-  //   };
-  //   return config;
-  // },
+  webpack(config) {
+    config.experiments = {
+      ...config.experiments,
+      topLevelAwait: true,
+    };
+    return config;
+  },
   experimental: {
     // turbo: {
     //     resolveAlias: {
